TCHECK-73: Ensure that gen/int returns a long
Commit 590cfae1 slightly changed the behavior of rand-range so that it
is no longer guaranteed to return an integer. Instead, it may return a
double if the lower and/or upper argument is a double. This directly
affects gen/int in cases where it is sized with a double.

Meanwhile, the recursive-gen function *always* uses a double for sizing
recursive generators, which means that inner generators will be sized
with a double, which means that gen/int will end up generating doubles.

This commit coerces the output of rand-range to long and adds
regression tests.